# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Lindéro API",
      "description": "The Lindéro project allows you to sell and buy paintings, sculptures and photographs. Sales are carried out through auction or at a fixed price.",
      "contact": {
        "name": "Elena Sheina",
        "email": "lena_sheina@bk.ru"
      },
      "version": "0.0.1"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Dev server"
      }
    ],
    "paths": {
...
      "/art": {
        "put": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Изменить данные товара",
          "operationId": "changeArt",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "required": [
                    "ArtChangeDTO",
                    "newPhotos"
                  ],
                  "type": "object",
                  "properties": {
                    "ArtChangeDTO": {
                      "$ref": "#/components/schemas/ArtChangeDTO"
                    },
                    "newPhotos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              },
              "application/octet-stream": {
                "schema": {
                  "required": [
                    "ArtChangeDTO",
                    "newPhotos"
                  ],
                  "type": "object",
                  "properties": {
                    "ArtChangeDTO": {
                      "$ref": "#/components/schemas/ArtChangeDTO"
                    },
                    "newPhotos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              },
              "application/json": {
                "schema": {
                  "required": [
                    "ArtChangeDTO",
                    "newPhotos"
                  ],
                  "type": "object",
                  "properties": {
                    "ArtChangeDTO": {
                      "$ref": "#/components/schemas/ArtChangeDTO"
                    },
                    "newPhotos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Данные успешно изменены",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            },
            "409": {
              "description": "Конфликтующее действие",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadActionException"
                  }
                }
              }
            },
            "400": {
              "description": "Неверные данные",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadCredentialsException"
                  }
                }
              }
            },
            "403": {
              "description": "Действие запрещено",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ForbiddenActionException"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Добавить товар",
          "operationId": "createArt",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "required": [
                    "ArtCreateDTO",
                    "photos"
                  ],
                  "type": "object",
                  "properties": {
                    "ArtCreateDTO": {
                      "$ref": "#/components/schemas/ArtCreateDTO"
                    },
                    "photos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              },
              "application/octet-stream": {
                "schema": {
                  "required": [
                    "ArtCreateDTO",
                    "photos"
                  ],
                  "type": "object",
                  "properties": {
                    "ArtCreateDTO": {
                      "$ref": "#/components/schemas/ArtCreateDTO"
                    },
                    "photos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              },
              "application/json": {
                "schema": {
                  "required": [
                    "ArtCreateDTO",
                    "photos"
                  ],
                  "type": "object",
                  "properties": {
                    "ArtCreateDTO": {
                      "$ref": "#/components/schemas/ArtCreateDTO"
                    },
                    "photos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "409": {
              "description": "Конфликтующее действие",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadActionException"
                  }
                }
              }
            },
            "400": {
              "description": "Неверные данные",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadCredentialsException"
                  }
                }
              }
            },
            "200": {
              "description": "Товар успешно добавлен",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Удалить товар",
          "operationId": "deleteArt",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntIdRequestDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Товар успешно удален",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            },
            "400": {
              "description": "Неверные данные",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadCredentialsException"
                  }
                }
              }
            },
            "403": {
              "description": "Действие запрещено",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ForbiddenActionException"
                  }
                }
              }
            }
          }
        }
      },
...
      "/{type}": {
        "get": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Получить все товары в зависимости от типа",
          "operationId": "getArts",
          "parameters": [
            {
              "name": "type",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Отправлен список товаров",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CommonArtDTO"
                  }
                }
              }
            },
            "400": {
              "description": "Неверные данные",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadCredentialsException"
                  }
                }
              }
            }
          }
        }
      },
      "/search/sculptures/object={input}": {
        "get": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Поиск скульптур по названию",
          "operationId": "searchSculptures",
          "parameters": [
            {
              "name": "input",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Отправлен список скульптур",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CommonArtDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/search/photos/object={input}": {
        "get": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Поиск фотографий по названию",
          "operationId": "searchPhotos",
          "parameters": [
            {
              "name": "input",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Отправлен список фотографий",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CommonArtDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/search/paintings/object={input}": {
        "get": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Поиск картин по названию",
          "operationId": "searchPaintings",
          "parameters": [
            {
              "name": "input",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Отправлен список картин",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CommonArtDTO"
                  }
                }
              }
            }
          }
        }
      },
...
      "/art/customer/{customerId}": {
        "get": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Получить все товары покупателя",
          "operationId": "getAllCustomerArts",
          "parameters": [
            {
              "name": "customerId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Неверные данные",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadCredentialsException"
                  }
                }
              }
            },
            "200": {
              "description": "Отправлен список товаров покупателя",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CustomerArtDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/art/artist/artistId={artistId}&currentId={currentId}": {
        "get": {
          "tags": [
            "Произведение искусства"
          ],
          "summary": "Получить все товары художника",
          "operationId": "getAllArtistArts",
          "parameters": [
            {
              "name": "artistId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            {
              "name": "currentId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Отправлен список товаров художника",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ArtistArtDTO"
                  }
                }
              }
            },
            "400": {
              "description": "Неверные данные",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/BadCredentialsException"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "ArtChangeDTO": {
          "type": "object",
          "properties": {
            "artId": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "deletePhotoUrls": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "changeMainPhoto": {
              "type": "boolean"
            },
            "isPrivate": {
              "type": "boolean"
            },
            "price": {
              "type": "number",
              "format": "double"
            },
            "description": {
              "type": "string"
            },
            "size": {
              "type": "string"
            },
            "frame": {
              "type": "boolean"
            },
            "tags": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "materials": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "BadActionException": {
          "type": "object",
          "properties": {
            "cause": {
              "type": "object",
              "properties": {
                "stackTrace": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "classLoaderName": {
                        "type": "string"
                      },
                      "moduleName": {
                        "type": "string"
                      },
                      "moduleVersion": {
                        "type": "string"
                      },
                      "methodName": {
                        "type": "string"
                      },
                      "fileName": {
                        "type": "string"
                      },
                      "lineNumber": {
                        "type": "integer",
                        "format": "int32"
                      },
                      "nativeMethod": {
                        "type": "boolean"
                      },
                      "className": {
                        "type": "string"
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                },
                "suppressed": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "stackTrace": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "classLoaderName": {
                              "type": "string"
                            },
                            "moduleName": {
                              "type": "string"
                            },
                            "moduleVersion": {
                              "type": "string"
                            },
                            "methodName": {
                              "type": "string"
                            },
                            "fileName": {
                              "type": "string"
                            },
                            "lineNumber": {
                              "type": "integer",
                              "format": "int32"
                            },
                            "nativeMethod": {
                              "type": "boolean"
                            },
                            "className": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "message": {
                        "type": "string"
                      },
                      "localizedMessage": {
                        "type": "string"
                      }
                    }
                  }
                },
                "localizedMessage": {
                  "type": "string"
                }
              }
            },
            "stackTrace": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "classLoaderName": {
                    "type": "string"
                  },
                  "moduleName": {
                    "type": "string"
                  },
                  "moduleVersion": {
                    "type": "string"
                  },
                  "methodName": {
                    "type": "string"
                  },
                  "fileName": {
                    "type": "string"
                  },
                  "lineNumber": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "nativeMethod": {
                    "type": "boolean"
                  },
                  "className": {
                    "type": "string"
                  }
                }
              }
            },
            "status": {
              "type": "string",
              "enum": [
                "100 CONTINUE",
                "101 SWITCHING_PROTOCOLS",
                "102 PROCESSING",
                "103 EARLY_HINTS",
                "103 CHECKPOINT",
                "200 OK",
                "201 CREATED",
                "202 ACCEPTED",
                "203 NON_AUTHORITATIVE_INFORMATION",
                "204 NO_CONTENT",
                "205 RESET_CONTENT",
                "206 PARTIAL_CONTENT",
                "207 MULTI_STATUS",
                "208 ALREADY_REPORTED",
                "226 IM_USED",
                "300 MULTIPLE_CHOICES",
                "301 MOVED_PERMANENTLY",
                "302 FOUND",
                "302 MOVED_TEMPORARILY",
                "303 SEE_OTHER",
                "304 NOT_MODIFIED",
                "305 USE_PROXY",
                "307 TEMPORARY_REDIRECT",
                "308 PERMANENT_REDIRECT",
                "400 BAD_REQUEST",
                "401 UNAUTHORIZED",
                "402 PAYMENT_REQUIRED",
                "403 FORBIDDEN",
                "404 NOT_FOUND",
                "405 METHOD_NOT_ALLOWED",
                "406 NOT_ACCEPTABLE",
                "407 PROXY_AUTHENTICATION_REQUIRED",
                "408 REQUEST_TIMEOUT",
                "409 CONFLICT",
                "410 GONE",
                "411 LENGTH_REQUIRED",
                "412 PRECONDITION_FAILED",
                "413 PAYLOAD_TOO_LARGE",
                "413 REQUEST_ENTITY_TOO_LARGE",
                "414 URI_TOO_LONG",
                "414 REQUEST_URI_TOO_LONG",
                "415 UNSUPPORTED_MEDIA_TYPE",
                "416 REQUESTED_RANGE_NOT_SATISFIABLE",
                "417 EXPECTATION_FAILED",
                "418 I_AM_A_TEAPOT",
                "419 INSUFFICIENT_SPACE_ON_RESOURCE",
                "420 METHOD_FAILURE",
                "421 DESTINATION_LOCKED",
                "422 UNPROCESSABLE_ENTITY",
                "423 LOCKED",
                "424 FAILED_DEPENDENCY",
                "425 TOO_EARLY",
                "426 UPGRADE_REQUIRED",
                "428 PRECONDITION_REQUIRED",
                "429 TOO_MANY_REQUESTS",
                "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
                "451 UNAVAILABLE_FOR_LEGAL_REASONS",
                "500 INTERNAL_SERVER_ERROR",
                "501 NOT_IMPLEMENTED",
                "502 BAD_GATEWAY",
                "503 SERVICE_UNAVAILABLE",
                "504 GATEWAY_TIMEOUT",
                "505 HTTP_VERSION_NOT_SUPPORTED",
                "506 VARIANT_ALSO_NEGOTIATES",
                "507 INSUFFICIENT_STORAGE",
                "508 LOOP_DETECTED",
                "509 BANDWIDTH_LIMIT_EXCEEDED",
                "510 NOT_EXTENDED",
                "511 NETWORK_AUTHENTICATION_REQUIRED"
              ]
            },
            "message": {
              "type": "string"
            },
            "suppressed": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "stackTrace": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "classLoaderName": {
                          "type": "string"
                        },
                        "moduleName": {
                          "type": "string"
                        },
                        "moduleVersion": {
                          "type": "string"
                        },
                        "methodName": {
                          "type": "string"
                        },
                        "fileName": {
                          "type": "string"
                        },
                        "lineNumber": {
                          "type": "integer",
                          "format": "int32"
                        },
                        "nativeMethod": {
                          "type": "boolean"
                        },
                        "className": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "message": {
                    "type": "string"
                  },
                  "localizedMessage": {
                    "type": "string"
                  }
                }
              }
            },
            "localizedMessage": {
              "type": "string"
            }
          }
        },
...
        "IntIdRequestDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
...
        "ArtCreateDTO": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "isPrivate": {
              "type": "boolean"
            },
            "price": {
              "type": "number",
              "format": "double"
            },
            "description": {
              "type": "string"
            },
            "size": {
              "type": "string"
            },
            "frame": {
              "type": "boolean"
            },
            "createDate": {
              "type": "string",
              "format": "date-time"
            },
            "tags": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "materials": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
...
      }
    }
  }